Add HTML template to prepare for React (#2)
* Add html loader and plugin

The HTML loader is used to pickup and translate HTML files. The plugin
then processes HTML templates to add webpack modules. This is required
for react so that a) the webpack bundle is added, b) React has an element
it can attach to, c) add web toolkits like Bootstrap.

* Add HTML template

This basic HTML template provides the root onto which react may attach
itself. It will be processed by webpack to insert the appropriate src
and link tags for JS and CSS respectively.

* Configure webpack to process HTML templates

html-loader is used to allow webpack to process HTML files. These
files are them processed by the HTML plugin to insert link and src
tags generated by Webpack.