Bootstrap react (#3)
* Configure webpack to use React 17 jsx transform

By default, the React preset does not use the new React 17 transform
for JSX. By setting the runtime to `automatic`, it will.

* Boostrap basic React application code

Wraps the app in React's strict mode fragment to ensure only the
best practises are followed. The App component itself is just a simple
boostrapping with a heading.