Fix multiple issues in webpack config
- In order to use the core bootstrap package easily, importing the CSS
directly is necessary with css-modules disabled. Bootstrap is a huge
pain in the ass with css-modules. If at some point it makes sense to
ditch sass, then this rule could be rewritten to differential between
"standard" css and css-modules by using the `.module.css` file
extension.
-  since sass files are currently being imported as css-modules, the
rule needs to enable css-modules in the configuration.
- JSX files are not resolved by webpack by default.