Basic layout (#9)
* Remove jquery and popper.js

React-bootstrap reimplements all of Bootstrap's JS so it should not be
necessary to have the core package's dependencies.

* Pretty up eslint config

* Fix multiple issues in webpack config

- In order to use the core bootstrap package easily, importing the CSS
directly is necessary with css-modules disabled. Bootstrap is a huge
pain in the ass with css-modules. If at some point it makes sense to
ditch sass, then this rule could be rewritten to differential between
"standard" css and css-modules by using the `.module.css` file
extension.
-  since sass files are currently being imported as css-modules, the
rule needs to enable css-modules in the configuration.
- JSX files are not resolved by webpack by default.

* Import bootstrap as a standard css file

By including the minified version by default, Bootstrap can be used as
normal by the app. Compiling the source may have some advantages but
the headache is not worth it at this time.

* Remove class from index

It doesn't make sense to include the class here. Since it is related
specifically to the app, it will be moved there.

* Add Header foundational element

* Reduce complexity of App

* Improve foundation layout

- Move Header component into its own directory
- Use an index for making imports cleaner
- Remove custom styling

* Tests should use .jsx exention

* Create Footer foundational component

* Add Footer to App